+++ src/main/java/org/bytedeco/javacpp/tools/InfoMap.java
@@ -42,7 +42,7 @@ public class InfoMap extends HashMap<String,List<Info>> {
     InfoMap parent = null;
     static final InfoMap defaults = new InfoMap(null)
         .put(new Info("basic/containers").cppTypes("std::array", "std::bitset", "std::deque", "std::list", "std::map", "std::queue", "std::set",
-                                                   "std::stack", "std::vector", "std::valarray", "std::pair", "std::forward_list",
+                                                   "std::stack", "std::vector", "std::valarray", "std::pair", "std::tuple", "std::forward_list",
                                                    "std::priority_queue", "std::unordered_map", "std::unordered_set", "std::optional", "std::variant"))
         .put(new Info("basic/types").cppTypes("signed", "unsigned", "char", "short", "int", "long", "bool", "float", "double",
                                               "_Bool", "_Complex", "_Imaginary", "complex", "imaginary"))